requirements.txt
python-dotenv
google-generativeai
pinecone
langchain
fastapi
uvicorn